requirements.txt
# Nota: este archivo define dependencias de runtime y desarrollo ligero,
# pero no sustituye a un lockfile resuelto por plataforma.
# Para reproducibilidad estricta hace falta generar un lockfile en un entorno valido.

# PyTorch con indice adicional para ruedas CUDA cuando proceda.
--extra-index-url https://download.pytorch.org/whl/cu128

# Base numerica y de datos
numpy>=1.24
pandas>=2.0
pyarrow>=15.0.0
pandas-datareader
PyYAML

# Descarga y procesamiento de mercado
yfinance
aiohttp

# Entrenamiento y forecasting
torch>=2.3.0
pytorch-lightning>=2.2.0
pytorch-forecasting>=1.0.0
optuna

# Visualizacion y aplicacion
streamlit
plotly
matplotlib
seaborn
nest_asyncio

# Utilidades opcionales presentes en el repositorio
fredapi
python-dotenv
